requirements.txt
python-docx
pdfplumber
nltk
spaCy
dateparser
flashtext
pandas
datetime
python-dateutil
rapidfuzz